+++ servers/lib/ws-handlerChain/managerChain.js
@@ -25,16 +25,14 @@ function handlerMsg4Str(message, connection) {
     const [ldap, service, fun] = message.utf8Data.split("->");
     console.log("%s: %s %s", ldap, service, fun);
     const [_, __] = fun.split(/(?:\()(.*)(?:\))/i); // 获取小括号中的内容
-    const _service = getService(service);
-    if (_service) {
+    const example = getService(service, _, {}); // 执行方法 传入要执行的service 方法名。 获取方法实例
+    if (example) {
       co(function* () {
-        if (_service[_]) {
-          const args = __? __.split("|"): null;
-          if (args) {
-            args.map((p,i) => args[i] = JSON.parse(p));
-          }
-          return yield _service[_].apply(null, args);
+        const args = __? __.split("|"): null;
+        if (args) {
+          args.map((p,i) => args[i] = JSON.parse(p));
         }
+        return yield example.apply(null, args);
       }).then(data => {
         connection.sendUTF(JSON.stringify({
           info: `${Date().toLocaleString()}: [SUCCESS] Receive msg`,
@@ -50,7 +48,7 @@ function handlerMsg4Str(message, connection) {
       });
     } else {
       connection.sendUTF(JSON.stringify({
-        info: `${Date().toLocaleString()}: [FAILED] exec func failed can't find this funcation`,
+        info: `${Date().toLocaleString()}: [FAILED] exec func failed can't find this funcation or you don't have permise to exec this func`,
       }));
     }
   } else {